+++ pkg/receive/tsdb.go
@@ -6,6 +6,7 @@ import (
 	"sync"
 
 	"github.com/go-kit/kit/log"
+	"github.com/go-kit/kit/log/level"
 	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/prometheus/storage/tsdb"
@@ -69,32 +70,32 @@ func (f *FlushableStorage) open() error {
 }
 
 // Flush temporarily stops the storage and flushes the WAL to blocks.
-// Note: this operation leaves the storage in the same state it was in.
+// Note: this operation leaves the storage closed.
 func (f *FlushableStorage) Flush() error {
+	_, err := os.Stat(filepath.Join(f.path, "wal"))
+	if os.IsNotExist(err) {
+		level.Info(f.l).Log("msg", "No WAL was found for flushing; ignoring.")
+		return nil
+	}
+	if err != nil {
+		return errors.Wrap(err, "stating WAL")
+	}
 	f.mu.Lock()
 	defer f.mu.Unlock()
-	var reopen bool
 	if !f.stopped {
 		if err := f.DB.Close(); err != nil {
 			return errors.Wrap(err, "stopping storage")
 		}
 		f.stopped = true
-		reopen = true
 	}
-	ro, err := promtsdb.OpenDBReadOnly(f.Dir(), f.l)
+	ro, err := promtsdb.OpenDBReadOnly(f.path, f.l)
 	if err != nil {
 		return errors.Wrap(err, "opening read-only DB")
 	}
-	if err := ro.FlushWAL(f.Dir()); err != nil {
+	if err := ro.FlushWAL(f.path); err != nil {
 		return errors.Wrap(err, "flushing WAL")
 	}
-	if err := os.RemoveAll(filepath.Join(f.Dir(), "wal")); err != nil {
-		return errors.Wrap(err, "removing stale WAL")
-	}
-	if reopen {
-		return errors.Wrap(f.open(), "re-starting storage")
-	}
-	return nil
+	return errors.Wrap(os.RemoveAll(filepath.Join(f.path, "wal")), "removing stale WAL")
 }
 
 // Close stops the storage.