+++ scripts/cfggen/main.go
@@ -15,6 +15,7 @@ import (
 	"github.com/thanos-io/thanos/pkg/objstore/azure"
 	"github.com/thanos-io/thanos/pkg/objstore/client"
 	"github.com/thanos-io/thanos/pkg/objstore/cos"
+	"github.com/thanos-io/thanos/pkg/objstore/filesystem"
 	"github.com/thanos-io/thanos/pkg/objstore/gcs"
 	"github.com/thanos-io/thanos/pkg/objstore/oss"
 	"github.com/thanos-io/thanos/pkg/objstore/s3"
@@ -30,12 +31,13 @@ import (
 
 var (
 	bucketConfigs = map[client.ObjProvider]interface{}{
-		client.AZURE:     azure.Config{},
-		client.GCS:       gcs.Config{},
-		client.S3:        s3.DefaultConfig,
-		client.SWIFT:     swift.SwiftConfig{},
-		client.COS:       cos.Config{},
-		client.ALIYUNOSS: oss.Config{},
+		client.AZURE:      azure.Config{},
+		client.GCS:        gcs.Config{},
+		client.S3:         s3.DefaultConfig,
+		client.SWIFT:      swift.SwiftConfig{},
+		client.COS:        cos.Config{},
+		client.ALIYUNOSS:  oss.Config{},
+		client.FILESYSTEM: filesystem.Config{},
 	}
 	tracingConfigs = map[trclient.TracingProvider]interface{}{
 		trclient.JAEGER:      jaeger.Config{},